Moves navigation logic from main.js to new component
Resovles #8189

Auditor: @bridiver
