+++ test/unit/app/renderer/components/navigation/navigatorTest.js
@@ -3,14 +3,14 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 /* global describe, before, after, it */
 
+require('../../../../braveUnit')
 const mockery = require('mockery')
 const {shallow} = require('enzyme')
 const assert = require('assert')
 const Immutable = require('immutable')
-let Main, NavigationBar
-require('../../braveUnit')
+let Navigator, NavigationBar
 
-describe('Main component unit tests', function () {
+describe('Navigator component unit tests', function () {
   before(function () {
     mockery.enable({
       warnOnReplace: false,
@@ -23,9 +23,9 @@ describe('Main component unit tests', function () {
     mockery.registerMock('../../extensions/brave/img/urlbar/browser_URL_fund_yes.svg', {})
     mockery.registerMock('../../extensions/brave/img/caret_down_grey.svg', 'caret_down_grey.svg')
     mockery.registerMock('../../extensions/brave/img/tabs/new_session.svg')
-    mockery.registerMock('electron', require('../../lib/fakeElectron'))
-    Main = require('../../../../js/components/main')
-    NavigationBar = require('../../../../app/renderer/components/navigation/navigationBar')
+    mockery.registerMock('electron', require('../../../../lib/fakeElectron'))
+    Navigator = require('../../../../../../app/renderer/components/navigation/navigator')
+    NavigationBar = require('../../../../../../app/renderer/components/navigation/navigationBar')
   })
 
   after(function () {
@@ -59,12 +59,32 @@ describe('Main component unit tests', function () {
     }]
   })
 
+  const customTitlebar = {
+    enabled: false,
+    captionButtonsVisible: false,
+    menubarVisible: false,
+    menubarTemplate: null,
+    menubarSelectedIndex: undefined,
+    contextMenuSelectedIndex: null,
+    lastFocusedSelector: undefined,
+    isMaximized: false
+  }
+
+  const activeTab = appState.getIn(['tabs', 0])
+
   describe('when user has history going forwards and backwards', function () {
     let wrapper
 
     before(function () {
       wrapper = shallow(
-        <Main windowState={windowState} appState={appState} />
+        <Navigator
+          windowState={windowState}
+          appState={appState}
+          activeTab={activeTab}
+          shouldAllowWindowDrag={false}
+          customTitlebar={customTitlebar}
+          activeSiteSettings={null}
+        />
       )
     })
 
@@ -95,7 +115,14 @@ describe('Main component unit tests', function () {
       })
 
       wrapper = shallow(
-        <Main windowState={windowState} appState={appState2} />
+        <Navigator
+          windowState={windowState}
+          appState={appState2}
+          activeTab={activeTab}
+          shouldAllowWindowDrag={false}
+          customTitlebar={customTitlebar}
+          activeSiteSettings={null}
+        />
       )
     })
 
@@ -129,7 +156,14 @@ describe('Main component unit tests', function () {
 
     before(function () {
       let wrapper = shallow(
-        <Main windowState={windowState} appState={appState} />
+        <Navigator
+          windowState={windowState}
+          appState={appState}
+          activeTab={activeTab}
+          shouldAllowWindowDrag={false}
+          customTitlebar={customTitlebar}
+          activeSiteSettings={null}
+        />
       )
       instance = wrapper.instance()
     })
